+++ RELEASES.md
@@ -40,6 +40,7 @@ Upcoming
 - Add configuration warning to use FixedQuerySender instead of DirectQuerySender
 - Install a Servlet Filter to protect the Struts 1 Servlet from ClassLoader Manipulation attacks
 - `(end of v6.1-RC1)`
+- Improve IBIS console function "Show Adapter Statistics"